Added saveInventory() function to main.cpp
The saveInventory() function was appended to main.cpp. The output of
this function is a file named inventory.inventory in the save directory
with one item name per line. Whatever the quantity of the item is in the
inventory is the number of times the item's name is output to this file
(so if you have 3 clubs, club is written 3 times to the
inventory.inventory file). This way it should be pretty simple to read
in the file to build your inventory when a saved game is loaded.