Debugging and updates to labors 1, 2, and 3
Debugged issue with improperly loading roomItems when loading a room.

Got labor 1 to nullify both trail and cave exits when boulder is pushed.

Similarly created event for labor 2 when hydra head is give to king.
This opens up path north from nemea to ceryneia. Also, modified long
desc, short desc, and visited (so long desc gets written next visit) on
nemea and both descriptions on throne room as well as updating the
current labor and the king talk response.

Added event for labor 3 when hind is given to king. This changes the
current labor, opens up the exit east from throne room to port of
tiryns, revises the short and long descriptions of throne room, and king
talk, and puts coin in Herc's pocket